+++ storage/saver_test.go
@@ -14,7 +14,7 @@ const (
 var (
 	testRecord = pb.Record{
 		Id:   666,
-		Data: []float32{0.6, 0.6, 0.6},
+		Data: []float64{0.6, 0.6, 0.6},
 		Meta: map[string]string{"666": "666"},
 	}
 )
