+++ wrapper/record_test.go
@@ -167,7 +167,7 @@ func TestWrappedRecordSetMetaWithInvalidId(t *testing.T) {
 }
 
 func TestWrappedRecordDot(t *testing.T) {
-	testRecord.Data = []float32{3, 6, 9}
+	testRecord.Data = []float64{3, 6, 9}
 	shouldBe := 126.0
 
 	a := WrapRecord(nil, &testRecord)
@@ -199,7 +199,7 @@ func TestWrappedRecordDotWithIncompatibleSizes(t *testing.T) {
 }
 
 func TestWrappedRecordMagnitude(t *testing.T) {
-	testRecord.Data = []float32{0, 0, 2}
+	testRecord.Data = []float64{0, 0, 2}
 	shouldBe := 2.0
 	a := WrapRecord(nil, &testRecord)
 	if mag := a.Magnitude(); mag != shouldBe {
@@ -214,8 +214,8 @@ func TestWrappedRecordMagnitudeWithNull(t *testing.T) {
 }
 
 func TestWrappedRecordCosine(t *testing.T) {
-	a := WrapRecord(nil, &pb.Record{Data: []float32{3, 6, 9}})
-	b := WrapRecord(nil, &pb.Record{Data: []float32{0, 0, 0}})
+	a := WrapRecord(nil, &pb.Record{Data: []float64{3, 6, 9}})
+	b := WrapRecord(nil, &pb.Record{Data: []float64{0, 0, 0}})
 	if cos := a.Cosine(b); cos != 0.0 {
 		t.Fatalf("cosine similarity should be %f, got %f", 0.0, cos)
 	}
