+++ wrapper/records_test.go
@@ -22,7 +22,7 @@ const (
 var (
 	testRecord = pb.Record{
 		Id:   666,
-		Data: []float32{3, 6, 9},
+		Data: []float64{3, 6, 9},
 		Meta: map[string]string{
 			"foo":    "bar",
 			"some":   "thing",
@@ -31,7 +31,7 @@ var (
 	}
 	testShorterRecord = pb.Record{
 		Id:   777,
-		Data: []float32{1},
+		Data: []float64{1},
 		Meta: map[string]string{},
 	}
 )
